+++ MyMusic/DataAccess/FanDataAccess/clsFanDA.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DataAccess
+namespace DataAccess.FanDataAccess
 {
     public class clsFanDA
     {
