+++ MyMusic/DataAccess/UserDataAccess/clsUserDA.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DataAccess
+namespace DataAccess.UserDataAccess
 {
     public class clsUserDA
     {
