+++ MyMusic/DataAccess/clsConnection.cs
@@ -20,7 +20,7 @@ namespace DataAccess
         {
             //conSettings = ConfigurationManager.ConnectionStrings["myMusicConnection"];
             //connectionString = conSettings.ConnectionString;
-            conn = new SqlConnection("Server=localhost;Database=myMusic;User ID=AdminMyMusic;Password=x100preXD;");
+            conn = new SqlConnection("Server=tcp:mymusic.database.windows.net,1433;Data Source=mymusic.database.windows.net;Initial Catalog=myMusic;Persist Security Info=False;User ID=David;Password=x100preXD;Pooling=False;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;");
         }
         public SqlConnection getPort()
         {
